Very dirty hack to clip draw line and image inside plot
Plot's drawing equivalents do clip, so I guess it's good for performance.
Not using the previous set_clip_space as it clips when any of the points
is outside the clip space, which is not what this is doing. The element
should be completly outside the space to be clipped.